+++ package.json
@@ -8,6 +8,7 @@
   "author": "Dat Team",
   "dependencies": {
     "base-elements": "^2.2.1",
+    "cache-element": "^3.0.0",
     "choo": "^5.1.0",
     "choo-log": "^6.0.0",
     "choo-persist": "^3.0.0",
@@ -22,10 +23,12 @@
     "electron-window": "^0.8.1",
     "envobj": "^1.0.2",
     "explain-error": "^1.0.3",
+    "microcomponent": "^1.0.0",
     "minimist": "^1.2.0",
     "mkdirp": "^0.5.1",
     "ms": "^0.7.2",
     "multidat": "^3.1.0",
+    "nanocomponent": "^3.0.2",
     "nanomorph": "^2.1.3",
     "prettier-bytes": "^1.0.3",
     "rimraf": "^2.5.4",
